pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mechforge-ai"
version = "0.4.0"
description = "MechForge AI - Mechanical Engineering AI Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "MechForge Team"}
]
keywords = ["mechanical engineering", "ai", "cad", "cae", "fem", "llm", "rag", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Mechanical Engineering",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "requests>=2.28",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typing-extensions>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
    "pre-commit>=3.0",
]
llm = [
    "llama-cpp-python>=0.2.0",
]
rag = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "rank-bm25>=0.2.2",
]
work = [
    "typer>=0.9.0",
    "gmsh>=4.13.0",
    "pyvista>=0.44.0",
    "numpy>=1.24.0",
]
work-full = [
    "mechforge-ai[work]",
    "pandas>=2.0.0",
]
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "psutil>=5.9.0",
]
web-full = [
    "mechforge-ai[web,work,rag]",
]
gui = [
    "PySide6>=6.6.0",
]
build = [
    "pyinstaller>=6.0.0",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
    "mkdocs-minify-plugin>=0.7.0",
    "mkdocs-git-revision-date-localized-plugin>=1.2.0",
    "pygments>=2.16.0",
    "pymdown-extensions>=10.3.0",
]
all = [
    "mechforge-ai[dev,llm,rag,work,web,gui,docs]"
]

[project.scripts]
mechforge = "mechforge_ai.terminal:main"
mechforge-k = "mechforge_knowledge.cli:main"
mechforge-work = "mechforge_work.work_cli:main"
mechforge-web = "mechforge_web.main:main"
mechforge-model = "mechforge_ai.model_cli:main"
mechforge-gui = "mechforge_gui_ai.app:main"

[project.urls]
Homepage = "https://github.com/yd5768365-hue/mechforge"
Documentation = "https://github.com/yd5768365-hue/mechforge#readme"
Repository = "https://github.com/yd5768365-hue/mechforge"
Issues = "https://github.com/yd5768365-hue/mechforge/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mechforge_*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
mechforge_core = ["*.yaml", "*.json"]
mechforge_theme = ["*.css", "*.html"]

# ========== Black 配置 ==========
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# ========== Ruff 配置 ==========
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # Line too long (handled by Black)
    "B008",  # FastAPI Depends in default args (standard FastAPI pattern)
    "N801",  # Class naming (log_performance is intentional decorator-style)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__ files
"examples/*" = ["N806", "E741"]  # Allow physics variable names in examples
"demo_*.py" = ["F401"]  # Allow import checks in demo files
"test_*.py" = ["F401", "E722", "F811"]  # Allow test-specific patterns
"run_*.py" = ["E402", "F401"]  # Allow script-style imports
"fix_*.py" = ["E741"]  # Allow legacy variable names in fix scripts

# ========== MyPy 配置 ==========
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "gmsh.*",
    "pyvista.*",
    "trimesh.*",
    "chromadb.*",
    "sentence_transformers.*",
    "llama_cpp.*",
]
ignore_missing_imports = true

# ========== Pytest 配置 ==========
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
